refactor(parse): use web ds file as 'proxy' instead of its own ds type
